requirements.txt
google-api-python-client==1.12.3
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-auth-httplib2==0.0.4
cakap==0.5.0
WeasyPrint==51
python-dotenv>=0.14.0
jinja2>=2.11.2
